Replace SimControls::nTrialRemain/trialDone with atomic startTrial
The old pair of calls (an atomic read, then a separate atomic
decrement) left a window between the two operations where another
thread's decrement could be missed, letting more than nTrial trials
run. startTrial() reads and decrements nTrialRemain_ in one atomic
capture, returning the trial number claimed (or 0 once none remain).

nTrialRemain_ is now a signed long rather than unsigned: once all
trials are claimed, concurrent callers that lose the race to decrement
first still read values <= 0 and correctly report no trials left,
rather than wrapping around to a huge unsigned value and running
indefinitely.

+++ src/io/SimControls.cpp
@@ -181,7 +181,7 @@ io::SimControls::SimControls(const toml::table& inputDeck)
     if (nTrialInput.has_value())
     {
         nTrial_ = nTrialInput.value();
-        nTrialRemain_ = nTrial_;
+        nTrialRemain_ = static_cast<long>(nTrial_);
     }
 
     // Handle output time generation
